reorganize code
